+++ src/components/admin/feeds/TopServiceSuggestions.jsx
@@ -15,6 +15,10 @@ import { avatarStyle } from "../../ResponsiveNav";
 import SeeMoreLink from "../../SeeMoreLink";
 import Username from "../../Username";
 import { PostCard, PostCardSkeleton } from "./DiscoverPostTabs";
+import {
+  useGetServicesFirstPage,
+  usePageinatedServices,
+} from "../../../hooks/useServices";
 
 const TopServiceSuggestions = () => {
   return (
@@ -29,11 +33,7 @@ const TopServiceSuggestions = () => {
 export default TopServiceSuggestions;
 
 export function TopServices() {
-  const { data: services, isLoading } = useQuery({
-    queryKey: ["services"],
-    queryFn: getServices,
-  });
-
+  const { data: services, isLoading } = useGetServicesFirstPage();
   const service = services?.[0];
 
   return (
@@ -87,7 +87,12 @@ export function Suggestions({
   );
 }
 
-export function SuggestionList({ hasSeeMore, associated = false, thisUser }) {
+export function SuggestionList({
+  hasSeeMore,
+  associated = false,
+  thisUser,
+  viewMoreUrl,
+}) {
   const { user: currentUser } = useAuth();
 
   const queryKey = [
@@ -159,8 +164,8 @@ export function SuggestionList({ hasSeeMore, associated = false, thisUser }) {
           })
         )}
       </ul>
-      {hasSeeMore && shownUsers?.length > 10 && (
-        <SeeMoreLink url="/representatives" />
+      {hasSeeMore && shownUsers?.length > 1 && (
+        <SeeMoreLink url={viewMoreUrl} />
       )}
     </section>
   );