+++ changes.md
@@ -1,2 +1,7 @@
-- Add image preview and blur to messages
-- messages now use a single socket
+- corrected some pagination issues with products and services
+- fixed image scaling of listed products in products page
+- Representatives invitation now works
+
+- There now a view more button in the People Associated sections for user and company profile that leads to the representative page
+- Representatives page is now paginated
+- Representatives page can now load only a specific companies representativea